Split signup and signin testing
